scope: SC-42 adopt orphaned gnn/testing tests into tests/testing
- git mv test_comprehensive.py, test_integration.py, test_xml_parser_only.py
  from src/gnn/testing/ to tests/testing/ (now inside pytest testpaths).
- Strip obsolete src-path sys.path hacks; use the gnn.* canonical import
  surface like every other test module.
- test_round_trip.py stays in src as the production harness surface:
  gnn.schema_validator.validator lazily imports GNNRoundTripTester from it
  (validator.py:90); moving it would break the ROUND_TRIP validation level
  and the wheel install surface. Instead, extract the orphaned
  TestGNNRoundTrip TestCase into tests/testing/test_round_trip_cases.py,
  importing the harness via its public gnn.testing.test_round_trip path.
- src/gnn/testing/test_round_trip.py drops the now-unused unittest import;
  harness classes and __main__ runner unchanged.
