+++ src/gnn/testing/test_round_trip.py
@@ -26,7 +26,6 @@ import os
 import re
 import sys
 import tempfile
-import unittest
 from dataclasses import dataclass, field
 from datetime import datetime
 from pathlib import Path
@@ -1167,83 +1166,6 @@ class GNNRoundTripTester(RoundTripComparisonMixin, RoundTripReportMixin):
         }
         return cast("str", extensions.get(format, "txt"))
 
-
-class TestGNNRoundTrip(unittest.TestCase):
-    """Unit tests for the round-trip testing system."""
-
-    def setUp(self) -> Any:
-        """Set up test environment."""
-        if not GNN_AVAILABLE:
-            self.skipTest("GNN module not available")
-
-        self.tester = GNNRoundTripTester()
-
-    def test_reference_file_exists(self) -> Any:
-        """Test that the reference file exists and is readable."""
-        self.assertTrue(
-            self.tester.reference_file.exists(),
-            f"Reference file not found: {self.tester.reference_file}",
-        )
-
-    def test_reference_file_validation(self) -> Any:
-        """Test that the reference file validates correctly."""
-        if self.tester.validator:
-            result = self.tester.validator.validate_file(self.tester.reference_file)
-            self.assertTrue(
-                result.is_valid, f"Reference file validation failed: {result.errors}"
-            )
-        else:
-            self.skipTest("Validator not available")
-
-    def test_comprehensive_round_trip(self) -> Any:
-        """Test comprehensive round-trip conversion."""
-        report = self.tester.run_comprehensive_tests()
-
-        # Basic assertions
-        self.assertGreater(report.total_tests, 0, "No tests were run")
-        self.assertGreaterEqual(report.successful_tests, 0, "No successful tests")
-
-        # Generate report
-        report_content = self.tester.generate_report(report)
-        self.assertIn("GNN Round-Trip Testing Report", report_content)
-
-        # Log results
-        print(
-            f"\nRound-trip test results: {report.successful_tests}/{report.total_tests} passed"
-        )
-        print(f"Success rate: {report.get_success_rate():.1f}%")
-
-        if report.failed_tests > 0:
-            print("Failed formats:")
-            for result in report.round_trip_results:
-                if not result.success:
-                    print(f"  - {result.target_format.value}: {result.errors}")
-
-    def test_specific_format_round_trip(self) -> Any:
-        """Test round-trip for a specific format (JSON)."""
-        if not self.tester.parsing_system:
-            self.skipTest("Parsing system not available")
-
-        # Parse reference
-        reference_result = self.tester.parsing_system.parse_file(
-            self.tester.reference_file, GNNFormat.MARKDOWN
-        )
-        self.assertTrue(reference_result.success, "Failed to parse reference file")
-
-        # Test JSON round-trip
-        json_result = self.tester._test_round_trip(
-            reference_result.model, GNNFormat.JSON
-        )
-
-        if not json_result.success:
-            print("JSON round-trip failed:")
-            print(f"  Errors: {json_result.errors}")
-            print(f"  Differences: {json_result.differences}")
-
-        # Should succeed for JSON format
-        self.assertTrue(
-            json_result.success, f"JSON round-trip failed: {json_result.errors}"
-        )
 
 
 if __name__ == "__main__":
