+++ tests/testing/test_xml_parser_only.py
@@ -4,12 +4,8 @@ Test to isolate the XML parser duplication issue.
 """
 
 import json
-import sys
-from pathlib import Path
 from typing import Any
 
-# Add src directory to path
-sys.path.insert(0, str(Path(__file__).parent.parent.parent))
 
 from gnn.parsers.xml_parser import XMLGNNParser
